+++ integrations/jira/initialize_client.py
@@ -1,0 +1,11 @@
+from jira.client import JiraClient
+from port_ocean.context.ocean import ocean
+
+
+def create_jira_client() -> JiraClient:
+    """Create JiraClient with current configuration."""
+    return JiraClient(
+        ocean.integration_config["jira_host"],
+        ocean.integration_config["atlassian_user_email"],
+        ocean.integration_config["atlassian_user_token"],
+    )