+++ port_ocean/clients/port/mixins/integrations.py
@@ -13,6 +13,7 @@ if TYPE_CHECKING:
     from port_ocean.core.handlers.port_app_config.models import PortAppConfig
 
 
+ORG_USE_PROVISIONED_DEFAULTS_FEATURE_FLAG = "USE_PROVISIONED_DEFAULTS"
 INTEGRATION_POLLING_INTERVAL_INITIAL_SECONDS = 3
 INTEGRATION_POLLING_INTERVAL_BACKOFF_FACTOR = 1.55
 INTEGRATION_POLLING_RETRY_LIMIT = 30
@@ -68,21 +69,26 @@ class IntegrationClientMixin:
         return response
 
     async def get_current_integration(
-        self, should_raise: bool = True, should_log: bool = True
+        self,
+        should_raise: bool = True,
+        should_log: bool = True,
+        has_provision_feature_flag: bool = False,
     ) -> dict[str, Any]:
         response = await self._get_current_integration()
         handle_status_code(response, should_raise, should_log)
         integration = response.json().get("integration", {})
         if integration.get("config", None) or not integration:
             return integration
-        is_provision_enabled_for_integration = integration.get(
-            "installationAppType", None
-        ) and (
-            await self.is_integration_provision_enabled(
-                integration.get("installationAppType", ""),
-                should_raise,
-                should_log,
+        is_provision_enabled_for_integration = (
+            integration.get("installationAppType", None)
+            and (
+                await self.is_integration_provision_enabled(
+                    integration.get("installationAppType", ""),
+                    should_raise,
+                    should_log,
+                )
             )
+            and has_provision_feature_flag
         )
 
         if is_provision_enabled_for_integration: