+++ integrations/gitlab/CHANGELOG.md
@@ -7,6 +7,22 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 <!-- towncrier release notes start -->
 
+0.2.41 (2025-02-23)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.20.2
+
+
+0.2.40 (2025-02-23)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.20.1
+
+
 0.2.39 (2025-02-19)
 ===================